* fixing install script

+++ install-mylauncher.sh
@@ -4,16 +4,16 @@ homedir="/home/cpi"
 
 if [ ! -d $homedir ]; then
   echo "No homedir [$homedir] found, don't know what to do."
-  return
-elif [ -f $homedir/.start ]; then
+  exit 1
+elif [ -f $homedir/.startrc ]; then
   echo "It seem that switch launcher script is already installed"
-  return
+  exit 1
 elif [ -d $homedir/mylauncher ]; then
   echo "$homedir/mylauncher folder already exists, rename or remove it first if you want to install fresh copy of mylauncher"
-  return
+  exit 1
 fi
 
-whiptail --title "Confirm" --yesno "This script will install \"mylauncher\" in $homedir folder. Continue?" $ALERT_HEIGHT $WIDTH
+whiptail --title "Confirm" --yesno "This script will install \"mylauncher\" in $homedir folder. Continue?" 8 40
 x=$?
 
 if ! [ $x -eq 0 ]; then
@@ -27,7 +27,7 @@ cd "$homedir"
 wget https://github.com/domichal/mylauncher/archive/master.zip && unzip master.zip && rm master.zip
 mv mylauncher-master mylauncher
 
-cat <<EOT>> .start
+cat <<EOT>> .startrc
 #!/bin/bash
 
 LAUNCHER=mylauncher
@@ -37,16 +37,18 @@ if [ -f "$homedir/\$LAUNCHER/.cpirc" ] && [ -z "\$SSH_CLIENT" ] && [ -z "\$SSH_T
     . "$homedir/\$LAUNCHER/.cpirc"
 fi
 EOT
-chmod 644 .start
+chmod 644 .startrc
 
-if [ "$(cat .bashrc | grep '\.start')" != "" ]; then
-    echo "Start entry has already been added to $home/.bashrc"
+if [ "$(cat .bashrc | grep '\.startrc')" != "" ]; then
+  echo "Start entry has already been added to $home/.bashrc"
 else
-    cp -f .bashrc .bashrc.bak
-    sed -i '/^if \[ -f .*\.cpirc/i \if [ -f ~/.start ]; then\n    . ~/.start\nfi\n' .bashrc || echo -e "Something went wrong when trying to modify $homedir/.bash\nPlease edit it manually, more info here: https://github.com/domichal/mylauncher/blob/dev/README.md#installation"
+  cp -f .bashrc .bashrc.bak
+  sed -i "/^if \[ -f .*\.cpirc/i \if [ -f $homedir/.startrc ]; then\n    . $homedir/.startrc\nel" .bashrc || echo -e "Something went wrong when trying to modify $homedir/.bash\nPlease edit it manually, more info here: https://github.com/domichal/mylauncher/blob/dev/README.md#installation"
 fi
 echo "Installation finished"
 
 cd $pwd
-rm -- "$0"
+if [ "$0" != bash ]; then
+  rm -- "$0"
+fi
 exit 0
